spike: move createMissingLinkedNotes to cli
